+++ src/de/bielefeld/umweltamt/aui/module/objektpanels/Anh49DetailsPanel.java
@@ -107,7 +107,6 @@ import java.awt.event.KeyEvent;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
 import java.util.ArrayList;
-import java.util.List;
 
 import javax.swing.AbstractAction;
 import javax.swing.Action;
@@ -126,15 +125,14 @@ import com.jgoodies.forms.layout.CellConstraints;
 import com.jgoodies.forms.layout.FormLayout;
 
 import de.bielefeld.umweltamt.aui.GUIManager;
-import de.bielefeld.umweltamt.aui.mappings.DatabaseQuery;
 import de.bielefeld.umweltamt.aui.mappings.indeinl.Anh49Abscheiderdetails;
 import de.bielefeld.umweltamt.aui.mappings.indeinl.Anh49Fachdaten;
 import de.bielefeld.umweltamt.aui.module.BasisObjektBearbeiten;
 import de.bielefeld.umweltamt.aui.module.common.editors.AbscheiderEditor;
+import de.bielefeld.umweltamt.aui.module.common.tablemodels.Anh49AbscheiderModel;
 import de.bielefeld.umweltamt.aui.utils.AuikLogger;
 import de.bielefeld.umweltamt.aui.utils.TabAction;
 import de.bielefeld.umweltamt.aui.utils.TableFocusListener;
-import de.bielefeld.umweltamt.aui.utils.tablemodelbase.ListTableModel;
 
 /**
  * Das Abscheiderdetails-Tab beim Objekt bearbeiten
@@ -149,89 +147,6 @@ public class Anh49DetailsPanel extends JPanel {
      */
     /** Logging */
     private static final AuikLogger log = AuikLogger.getLogger();
-
-    private class Anh49AbscheiderModel extends ListTableModel {
-        private static final long serialVersionUID = 6154019963876247085L;
-        private Anh49Fachdaten fachdaten;
-
-        /**
-         * Erzeugt ein neues Abscheider-TableModel. Dieses hat die Spalten
-         * "Abscheider", "Von", "Lage" und "Bemerkung".
-         */
-        public Anh49AbscheiderModel() {
-            super(new String[] {"Abscheider", "Von", "Lage", "Nenngröße", "Hersteller", "Bemerkung"},
-                false, true);
-        }
-
-        /**
-         * Setzt das Fachdatenobjekt, nach dessen Abscheider-Details gesucht
-         * werden soll.
-         * @param fachdaten Das Anhang49-Fachdatenobjekt
-         */
-        public void setFachdaten(Anh49Fachdaten fachdaten) {
-            this.fachdaten = fachdaten;
-            updateList();
-        }
-
-        @Override
-        public Object getColumnValue(Object objectAtRow, int columnIndex) {
-            Anh49Abscheiderdetails details = (Anh49Abscheiderdetails) objectAtRow;
-
-            Object tmp;
-
-            switch (columnIndex) {
-                case 0:
-                    tmp = details.getAbscheidernr();
-                    break;
-                case 1:
-                    tmp = details.getVon();
-                    break;
-                case 2:
-                    tmp = details.getLage();
-                    break;
-                case 3:
-                    tmp = details.getNenngroesse();
-                    break;
-                case 4:
-                    tmp = details.getHersteller();
-                    break;
-                case 5:
-                    tmp = details.getBemerkung();
-                    break;
-                default:
-                    tmp = null;
-            }
-
-            return tmp;
-        }
-
-        @Override
-        public boolean objectRemoved(Object objectAtRow) {
-            Anh49Abscheiderdetails removedAbsch = (Anh49Abscheiderdetails) objectAtRow;
-            boolean removed;
-
-            if (removedAbsch.getId() != null) {
-                removed = Anh49Abscheiderdetails.delete(removedAbsch);
-            } else {
-                removed = true;
-            }
-
-            return removed;
-        }
-
-        @Override
-        public void updateList() {
-            if (fachdaten != null) {
-            	List<Anh49Abscheiderdetails>  x = DatabaseQuery.getAbscheiderDetails(fachdaten);
-                setList(x);
-            }
-            fireTableDataChanged();
-        }
-
-        public Anh49Abscheiderdetails getRow(int rowIndex) {
-            return (Anh49Abscheiderdetails) getObjectAtRow(rowIndex);
-        }
-    }
 
     private String name;
